src/main/java/umu/tds/AppMusic/controlador/RepositorioUsuarios.java
package umu.tds.AppMusic.controlador;

import java.util.HashMap;
import java.util.LinkedList;
import java.util.List;

import umu.tds.AppMusic.dao.DAOException;
import umu.tds.AppMusic.dao.FactoriaDao;
import umu.tds.AppMusic.modelo.Usuario;

public enum RepositorioUsuarios {
	INSTANCE;

	private FactoriaDao factoria;

	private HashMap<Integer, Usuario> usuariosPorID;
	private HashMap<String, Usuario> usuariosPorLogin;

	private RepositorioUsuarios() {
		usuariosPorID = new HashMap<Integer, Usuario>();
		usuariosPorLogin = new HashMap<String, Usuario>();

		try {
			factoria = FactoriaDao.getInstancia();

			List<Usuario> listausuarios = factoria.getUsuarioDAO().obtenerTodosLosUsuarios();
			for (Usuario usuario : listausuarios) {
				usuariosPorID.put(usuario.getId(), usuario);
				usuariosPorLogin.put(usuario.getLogin(), usuario);
			}
		} catch (DAOException eDAO) {
			eDAO.printStackTrace();
		}
	}

	public List<Usuario> findUsuarios() throws DAOException {
		return new LinkedList<Usuario>(usuariosPorLogin.values());
	}

	public Usuario findUsuario(String login) {
		if (login == null) {
			return null;
		}
		return usuariosPorLogin.get(login);
	}

	public Usuario findUsuario(int id) {
		return usuariosPorID.get(id);
	}

	public void addUsuario(Usuario usuario) {
		usuariosPorID.put(usuario.getId(), usuario);
		usuariosPorLogin.put(usuario.getLogin(), usuario);
	}

	public void removeUsuario(Usuario usuario) {
		usuariosPorID.remove(usuario.getId());
		usuariosPorLogin.remove(usuario.getLogin());
	}

}
